move float params out of smooth.rs
